Conflicts:
	tools/all/lib/minko.project.lua

Former-commit-id: d3bc56935c4462b47fcb3238644cf87ac0d67e9b
Former-commit-id: 98684ee8d902e9eceb1869a18f0356d93df32267

+++ plugins/mk/src/minko/deserialize/TypeDeserializer.cpp
@@ -111,7 +111,7 @@ TypeDeserializer::number(std::map<int, render::Texture::Ptr>&	idToTexture,
 		if (object["value"].type() == typeid(float))
 			material->set(nameConverter->convertString(propertyName), Any::cast<float>(object["value"]));
 		else if (object["value"].type() == typeid(int))
-			material->set(nameConverter->convertString(propertyName), Any::cast<int>(object["value"]));
+			material->set(nameConverter->convertString(propertyName), static_cast<float>(Any::cast<int>(object["value"])));
 		else
 			material->set(nameConverter->convertString(propertyName), Any::cast<unsigned int>(object["value"]));
